Inserindo stock

+++ model/stock.go
@@ -1,7 +1,7 @@
 package model
 
 type Stock struct {
-	ID     uint64 `gorm:"primary_key,autoIncrement"`
+	ID     int `gorm:"primary_key,autoIncrement"`
 	Ticker string
 	Price  float64
 }
